% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsano.R
\docType{package}
\name{ggsano}
\alias{ggsano}
\title{ggsano: A package for colours and templates based on the official Sciensano style guide.}
\description{
The ggsano package currently only provides the official Sciensano colour scales.
}
