% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_scien_d.R
\name{pal_scien_d}
\alias{pal_scien_d}
\title{Provide discrete Sciensano Colours}
\usage{
pal_scien_d(n = 13)
}
\arguments{
\item{n}{Number of colours that should be returned}
}
\value{
Character vector with #HEX colour codes
}
\description{
XXXXXXXXX
}
