% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_sciensano.R
\name{pal_sciensano}
\alias{pal_sciensano}
\title{Sciensano Color Palettes}
\usage{
pal_sciensano(alpha = 1)
}
\arguments{
\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Sciensano Color Palettes
}
\examples{
library("scales")
show_col(pal_sciensano()(10))
show_col(pal_sciensano(alpha = 0.6)(10))
}
\author{
Robby De Pauw <\email{Robby.DePauw@sciensano.be}>
}
