% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_sciensano_c.R
\name{pal_sciensano_c}
\alias{pal_sciensano_c}
\title{Provide Continuous Sciensano Colours}
\usage{
pal_sciensano_c(
  ...,
  low = "#B2D235",
  high = "#006838",
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour"
)
}
\arguments{
\item{low, high}{Colours for low and high ends of the gradient.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" - other values are deprecated.}

\item{na.value}{Colour to use for missing values}

\item{aesthetics}{Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with. This can be useful, for example, to apply colour settings to the colour and fill aesthetics at the same time, via aesthetics = c("colour", "fill").}
}
\value{
Character vector with #HEX colour codes
}
\description{
XXXXXXXXX
}
