% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_vizz.R
\name{pal_vizz}
\alias{pal_vizz}
\title{Vizualize Sciensano colours}
\usage{
pal_vizz(colours, labels = TRUE, borders = 1, cex_label = 1, ncol = NULL)
}
\arguments{
\item{colours}{A character vector of colours}

\item{labels}{Label each colour with its hex name?}

\item{borders}{Border colour for each tile. Default uses par("fg"). Use border = NA to omit borders.}

\item{ncol}{Size of printed labels, as multiplier of default size.}

\item{cex_labels}{Size of printed labels, as multiplier of default size.}
}
\value{
Number of columns. If not supplied, tries to be as square as possible.
}
\description{
A quick and dirty way to show colours in a plot.
}
