% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-helpers.R
\name{scale_color_sciensano_c}
\alias{scale_color_sciensano_c}
\alias{scale_colour_sciensano_c}
\alias{scale_fill_sciensano_c}
\title{Sciensano Continuous Color Scales}
\usage{
scale_color_sciensano_c(
  palette = c("red", "green", "red-yellow-green"),
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_colour_sciensano_c(
  palette = c("red", "green", "red-yellow-green"),
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_fill_sciensano_c(
  palette = c("red", "green", "red-yellow-green"),
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette type.
There are 3 available options:
\itemize{
\item \code{"red"}
\item \code{"green"}
\item \code{"red-yellow-green"}}}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{Logical. Should the order of the colors be reversed?}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_sciensano_c}} for details.
}
\examples{
\dontrun{library("ggplot2")
library("reshape2")
data("mtcars")

cor <- abs(cor(mtcars))
cor_melt <- melt(cor)

ggplot(
  cor_melt,
  aes(x = Var1, y = Var2, fill = value)
) +
  geom_tile(colour = "black", size = 0.3) +
  theme_bw() + scale_fill_sciensano_c("red-yellow-green")}
}
