% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{Database with official Sciensano Colors}
\format{
A data frame with official Sciensano colours:
\describe{
\item{all}{All colours (n = 13)}
\item{primary}{Primlary colours (4)}
\item{secondary}{Secondary colours (n = 9)}
\item{yellowaccent}{Colours with a yellow - red accent (n = 4)}
...
}
}
\source{
\url{http://www.sciensano.be/}
}
\usage{
x
}
\description{
Database with official Sciensano Colors
}
\keyword{datasets}
