% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsano.R
\docType{package}
\name{ggsano}
\alias{ggsano}
\title{ggsano: A package that implements the Sciensano house style in R}
\description{
The ggsano package currently provides three categories of important functions:
sciensano_style, sciensano_colours and sciensano_templates.
}
\section{sciensano_style functions}{

The sciensano_style functions can handle a ggplot2-object and recompile it
according to the Sciensano house style.
}

\section{sciensano_colours functions}{

The sciensano_colours functions return the Sciensano house style colours in
HEX format ('#XXXXXX'). Both discrete and continuous scales are available.
}

\section{sciensano_templates}{

TBD.
}

