% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-helpers.R
\name{scale_color_sciensano}
\alias{scale_color_sciensano}
\alias{scale_colour_sciensano}
\alias{scale_fill_sciensano}
\title{Sciensano Color Color Scales}
\usage{
scale_color_sciensano(palette = c("default"), alpha = 1, ...)

scale_colour_sciensano(palette = c("default"), alpha = 1, ...)

scale_fill_sciensano(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(14-color palette inspired by \emph{Sciensano}).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_sciensano}} for details.
}
\examples{
\dontrun{library("ggplot2")
data("diamonds")

ggplot(
  subset(diamonds, carat >= 2.2),
  aes(x = table, y = price, colour = cut)
) +
  geom_point(alpha = 0.7) +
  geom_smooth(method = "loess", alpha = 0.1, size = 1, span = 1) +
  theme_bw() + scale_color_sciensano()

ggplot(
  subset(diamonds, carat > 2.2 & depth > 55 & depth < 70),
  aes(x = depth, fill = cut)
) +
  geom_histogram(colour = "black", binwidth = 1, position = "dodge") +
  theme_bw() + scale_fill_sciensano()}
}
